+++ Flow/Changes.md
@@ -47,6 +47,7 @@
   - The resulting model still utilizes the original model's property factory
 - **ModelLike**`.contains(String)` and `.containsNonEmpty(String)` now support lazily initialized properties 
   (e.g. those returned via the new model validation implementation)
+- `~==` implementation in **Value** is now more flexible and takes into account approximate equality in inner values as well 
 - The generic type parameter in **PropertyFactory** is no longer restricted 
 - `View.fixed(...).mapValue(...)` now produces a fixed view instead of a **Lazy**
 - `View(...).mapValue(...)` now produces a call-by-name (mapping) view instead of a **Lazy**